Move starters file into a directory
